load items from file

+++ src/Types/Item.fs
@@ -4,5 +4,5 @@ open System
 type item = {
     Guid:Guid;
     Name:string;
-    Price:float;
+    Price:Decimal;
 }
